implement an opinionated format command
add a new fmt command that allows for formatting and checking the format of .hcl files.
In the process convert any key value pairs inside object literals to always
consistently use = instead of :.

It uses the std hclformat package for formatting.

A flag is present to turn off the key value normalizer. It may be removed once we
have enough experience with the formatter in the wild.

Signed-off-by: gotwarlost <krishnan.anantheswaran@elastic.co>